+++ tw-base-utils/src/test/java/com/transferwise/common/baseutils/UuidUtilsTest.java
@@ -34,15 +34,42 @@ class UuidUtilsTest extends BaseTest {
 
     long previousTime = -1;
     for (int i = 0; i < n; i++) {
+      System.out.println("Testing " + uuids[i]);
+      assertEquals(4, uuids[i].version());
       long time = uuids[i].getMostSignificantBits() >>> (64 - 38);
 
       if (previousTime != -1) {
         assertTrue(previousTime < time);
       }
       previousTime = time;
+    }
 
-      System.out.println(uuids[i]);
-      assertEquals(4, uuids[i].version());
+    assertTrue(Ordering.natural().isOrdered(Arrays.asList(uuids)));
+    assertEquals(n, Set.of(uuids).size());
+  }
+
+  @Test
+  void uuidv7IsGrowingOverTime() {
+    TestClock clock = TestClock.createAndRegister();
+
+    int n = 100;
+
+    UUID[] uuids = new UUID[n];
+    for (int i = 0; i < n; i++) {
+      uuids[i] = UuidUtils.generateUuidv7();
+      clock.tick(Duration.ofMillis(2));
+    }
+
+    long previousTime = -1;
+    for (int i = 0; i < n; i++) {
+      System.out.println("Testing " + uuids[i]);
+      assertEquals(7, uuids[i].version());
+      long time = uuids[i].getMostSignificantBits() >>> (64 - 48);
+
+      if (previousTime != -1) {
+        assertTrue(previousTime < time);
+      }
+      previousTime = time;
     }
 
     assertTrue(Ordering.natural().isOrdered(Arrays.asList(uuids)));