+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2025-07-10
+
+### Changed
+- Added support for generating [UUIDv7](https://uuid7.com/)s.
+- Deprecated prefix combined UUIDv4s in favour of UUIDv7.
+
 ## [1.13.4] - 2025-06-17
 
 ### Changed